First pass at UserProfile revamp
