+++ Tabloid/Models/UserProfile.cs
@@ -9,8 +9,8 @@ namespace Tabloid.Models
     {
         public int Id { get; set; }
 
-        [StringLength(28, MinimumLength = 28)]
-        public string FirebaseUserId { get; set; }
+        //[StringLength(28, MinimumLength = 28)]
+        //public string FirebaseUserId { get; set; }
 
         [Required]
         [MaxLength(50)]
