+++ Tabloid/Repositories/IUserProfileRepository.cs
@@ -1,10 +1,10 @@
-﻿using Tabloid.Models;
+﻿//using Tabloid.Models;
 
-namespace Tabloid.Repositories
-{
-    public interface IUserProfileRepository
-    {
-        void Add(UserProfile userProfile);
-        UserProfile GetByFirebaseUserId(string firebaseUserId);
-    }
-}
+//namespace Tabloid.Repositories
+//{
+//    public interface IUserProfileRepository
+//    {
+//        void Add(UserProfile userProfile);
+//        UserProfile GetByFirebaseUserId(string firebaseUserId);
+//    }
+//}
