+++ Tabloid/Repositories/IUserRepository.cs
@@ -1,0 +1,10 @@
+﻿using Tabloid.Models;
+
+namespace Tabloid.Repositories
+{
+    public interface IUserRepository
+    {
+        void Add(UserProfile userProfile);
+        UserProfile GetByEmail(string email);
+    }
+}
