+++ Tabloid/Startup.cs
@@ -22,24 +22,25 @@ namespace Tabloid
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddTransient<IUserProfileRepository, UserProfileRepository>();
+            //services.AddTransient<IUserProfileRepository, UserProfileRepository>();
+            services.AddTransient<IUserRepository, UserRepository>();
 
-            var firebaseProjectId = Configuration.GetValue<string>("FirebaseProjectId");
-            var googleTokenUrl = $"https://securetoken.google.com/{firebaseProjectId}";
-            services
-                .AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-                .AddJwtBearer(options =>
-                {
-                    options.Authority = googleTokenUrl;
-                    options.TokenValidationParameters = new TokenValidationParameters
-                    {
-                        ValidateIssuer = true,
-                        ValidIssuer = googleTokenUrl,
-                        ValidateAudience = true,
-                        ValidAudience = firebaseProjectId,
-                        ValidateLifetime = true
-                    };
-                });
+            //var firebaseProjectId = Configuration.GetValue<string>("FirebaseProjectId");
+            //var googleTokenUrl = $"https://securetoken.google.com/{firebaseProjectId}";
+            //services
+            //    .AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
+            //    .AddJwtBearer(options =>
+            //    {
+            //        options.Authority = googleTokenUrl;
+            //        options.TokenValidationParameters = new TokenValidationParameters
+            //        {
+            //            ValidateIssuer = true,
+            //            ValidIssuer = googleTokenUrl,
+            //            ValidateAudience = true,
+            //            ValidAudience = firebaseProjectId,
+            //            ValidateLifetime = true
+            //        };
+            //    });
 
             services.AddControllers();
 
@@ -47,25 +48,25 @@ namespace Tabloid
             {
                 c.SwaggerDoc("v1", new OpenApiInfo { Title = "Tabloid", Version = "v1" });
 
-                var securitySchema = new OpenApiSecurityScheme
-                {
-                    Name = "Authorization",
-                    BearerFormat = "JWT",
-                    Description = "JWT Authorization header using the Bearer scheme.",
-                    Type = SecuritySchemeType.ApiKey,
-                    In = ParameterLocation.Header,
-                    Reference = new OpenApiReference
-                    {
-                        Id = "Bearer",
-                        Type = ReferenceType.SecurityScheme,
-                    }
-                };
+                //var securitySchema = new OpenApiSecurityScheme
+                //{
+                //    Name = "Authorization",
+                //    BearerFormat = "JWT",
+                //    Description = "JWT Authorization header using the Bearer scheme.",
+                //    Type = SecuritySchemeType.ApiKey,
+                //    In = ParameterLocation.Header,
+                //    Reference = new OpenApiReference
+                //    {
+                //        Id = "Bearer",
+                //        Type = ReferenceType.SecurityScheme,
+                //    }
+                //};
 
-                c.AddSecurityDefinition("Bearer", securitySchema);
-                c.AddSecurityRequirement(new OpenApiSecurityRequirement
-                {
-                    { securitySchema, new[] { "Bearer"} }
-                });
+                //c.AddSecurityDefinition("Bearer", securitySchema);
+                //c.AddSecurityRequirement(new OpenApiSecurityRequirement
+                //{
+                //    { securitySchema, new[] { "Bearer"} }
+                //});
             });
 
         }
